backend_2.0/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
pillow==10.3.0
python-jose==3.3.0
